requirements.txt
huggingface_hub[cli]
PyQt6
tqdm
requests
pyinstaller
Pillow